Create inversion_id.py
I need to figure out how to remove duplicates from kmer lists while maintaining order (set doesn't work). Then, I can run a comparison between inverted kmers sequence 1 [a] and sequence 2 kmers [a]. Based on the pattern of matches and not, I'd like to create a graph where inversions are in one colour, matches in another, and non-unique kmers are in a third colour.